Use single name for Tembo packages (#3)
Sometimes Ubuntu changes the name of a package between OS releases,
often by including the major version in the package name. To keep things
simpler for Tembo Cloud, always use a single name for each Ubuntu
version, and pass the Ubuntu package name as the second argument to
`repackage`. The Tembo package name must be the same name as a control
file in the `control` directory.

Also make `local` variables lowercase and document the script.