Repackage into a tarball instead of an .deb (#4)
Turns out that using `dpkg` to install a package won't work on a CNPG
PostgreSQL node. Even if it only installs files on the persistent
volume, it also tries to update files elsewhere on the file system,
which won't work, since it's immutable. Besides, packages have to be
installed by `root`, and extension installation and other management
tasks of the Tembo operator are performed as the `postgres` user.

So refactor `repackage` to installed copy the `.so` files and symlinks
into a tarball. Include the `copyright` file, too, and build a `digests`
file with SHA-512 digests. These changes simplifies `repackage` a fair
bit.

Also add `install_package`, which knows how to validate the digests of
and instal the `.so` files in a tarball created by `repackage`. Update
the `README` and the CI workflow to install newly-created packages, and
build a second package in CI.