Create package cf files and use Make for async (#5)
Add `*.cfg` files for each package to the `packages` directory. Each one
is the name of the package and its contents define the upstream packages
to build from, one each for Noble (Ubuntu 24.04) and Jammy (Ubuntu
22.204). Teach `repackage` to read the package name from the appropriate
file.

Add `Makefile` to automate building all packages, and use `-j$(nproc)`
to build them in parallel.

Replace the funky `arch` function with `dpkg --print-architecture` for
the reliable name of the architecture for Apt packages. Also teach
`repackage` to set local variables instead of globals.

Add `.gitignore` to ignore tarballs, unpacked tarballs, and other
detritus, and update the README.