Add tembox binary (#7)
Simple Rust app for installing extensions created by `repackage` and
published by the Repackge workflow. Runs only on Linux for amd64 and
arm64. It uses native Rust for all functionality except for digest
verification, for which it currently shells out ot `shasum512`.

Add the Tembox workflow and have it build and test (no tests so far, but
eventually) the binary on arm64 and amd64 Linux and also package it up
for release. It includes the `README.md`, CHANGELOG.md`, and
`LICENSE.md` files in the release tarballs.