Build packages in OS subdirectory (#8)
Otherwise there is no way to tell the binaries apart for each OS.
Determine the OS name by reading `/etc/lsb-release`: include it in the
`Makefile`, source it in the Bash scripts, and parse it in `tembox`. The
`Makefile` requires package targets for each OS name until/unless I can
figure out how to use a wildcard in a target name.