Add tembox.cfg and validate it on install (#9)
The new `tembox.cfg` file contains a simple list of key-value pairs for
the package name, version, OS, and architecture. Teach `install_package`
to validate that the name, OS, and architecture are correct, and to
install the file into `/var/lib/postgresql/data/tembox`.

Then do the same for `tembox`. Move its most of the work it does into
functions, including a new one that parses key/value pairs, and have the
loading of `/etc/lsb-release` use the same parser. Also fix a bug where
`tembox` installed libraries in a subdirectory named for the OS, but
should not have.

Rejigger the Tembox workflow to replace the use of cross and to build on
native AMD and ARM runners and to install a package to verify they work.
Add configs to the release profile in `Cargo.toml` to reduce the size of
the binary (from 3.1M to 1.5M on arm64). Fix the Rust setup to use the
proper toolchain and an appropriate cache key.