Add pre-commit configuration and workflow (#10)
Add `.pre-commit-config.yaml` and the `make lint` target to run
[pre-commit] on the project to lint the Rust code, documentation, YAML
format, and to link and format shell scripts. Also add the
`.git/hooks/pre-commit` target to install a hook to run pre-commit for
every commit, and the `debian-lint-depends` target to install
dependencies in the workflow. Also tweak the `Makefile` line `include
/etc/lsb-release` fail quietly, and add a target to update all Rust
dependencies.

Include fixes form the YAML and shell formatting, mostly blank space
changes.

Add some missing lines to the examples in the README, and add
instructions for using pre-commit.

  [pre-commit]: https://pre-commit.com